Split interpretation tests into modules
