+++ tests/unit/disable_jit/test_ground_rule_helpers.py
@@ -1,111 +1,7 @@
-# tests/unit/test_interpretation.py
 import pytest
-import math
 from unittest.mock import Mock, call
 
-# Single, consistent import style: import the module once
-import pyreason.scripts.interpretation.interpretation_fp as interpretation
-import pyreason.scripts.numba_wrapper.numba_types.label_type as label
-
-# Bind pure-Python callables (works even if Numba compiled elsewhere)
-_is_sat_edge = interpretation.is_satisfied_edge
-is_satisfied_edge = getattr(_is_sat_edge, "py_func", _is_sat_edge)
-
-_is_sat_node = interpretation.is_satisfied_node
-is_satisfied_node = getattr(_is_sat_node, "py_func", _is_sat_node)
-
-_get_q_edge_groundings = interpretation.get_qualified_edge_groundings
-get_qualified_edge_groundings = getattr(_get_q_edge_groundings, "py_func", _get_q_edge_groundings)
-
-_get_q_node_groundings = interpretation.get_qualified_node_groundings
-get_qualified_node_groundings = getattr(_get_q_node_groundings, "py_func", _get_q_node_groundings)
-
-_get_rule_node_clause_grounding = interpretation.get_rule_node_clause_grounding
-get_rule_node_clause_grounding = getattr(_get_rule_node_clause_grounding, "py_func", _get_rule_node_clause_grounding)
-
-_get_rule_edge_clause_grounding = interpretation.get_rule_edge_clause_grounding
-get_rule_edge_clause_grounding = getattr(_get_rule_edge_clause_grounding, "py_func", _get_rule_edge_clause_grounding)
-
-_satisfies_threshold = interpretation._satisfies_threshold
-satisfies_threshold = getattr(_satisfies_threshold, "py_func", _satisfies_threshold)
-
-_check_node_thresh = interpretation.check_node_grounding_threshold_satisfaction
-check_node_grounding_threshold_satisfaction = getattr(_check_node_thresh, "py_func", _check_node_thresh)
-
-_check_edge_thresh = interpretation.check_edge_grounding_threshold_satisfaction
-check_edge_grounding_threshold_satisfaction = getattr(_check_edge_thresh, "py_func", _check_edge_thresh)
-
-_refine_groundings = interpretation.refine_groundings
-refine_groundings = getattr(_refine_groundings, "py_func", _refine_groundings)
-
-_check_all = interpretation.check_all_clause_satisfaction
-check_all_clause_satisfaction = getattr(_check_all, "py_func", _check_all)
-
-_add_node = interpretation._add_node
-add_node = getattr(_add_node, "py_func", _add_node)
-
-_add_edge = interpretation._add_edge
-add_edge = getattr(_add_edge, "py_func", _add_edge)
-
-_ground_rule = interpretation._ground_rule
-ground_rule = getattr(_ground_rule, "py_func", _ground_rule)
-
-_update_rule_trace = interpretation._update_rule_trace
-update_rule_trace = getattr(_update_rule_trace, "py_func", _update_rule_trace)
-
-_are_sat_node = interpretation.are_satisfied_node
-are_satisfied_node = getattr(_are_sat_node, "py_func", _are_sat_node)
-
-_are_sat_edge = interpretation.are_satisfied_edge
-are_satisfied_edge = getattr(_are_sat_edge, "py_func", _are_sat_edge)
-
-_is_sat_node_cmp = interpretation.is_satisfied_node_comparison
-is_satisfied_node_comparison = getattr(_is_sat_node_cmp, "py_func", _is_sat_node_cmp)
-
-_is_sat_edge_cmp = interpretation.is_satisfied_edge_comparison
-is_satisfied_edge_comparison = getattr(_is_sat_edge_cmp, "py_func", _is_sat_edge_cmp)
-
-_check_cons_node = interpretation.check_consistent_node
-check_consistent_node = getattr(_check_cons_node, "py_func", _check_cons_node)
-
-_check_cons_edge = interpretation.check_consistent_edge
-check_consistent_edge = getattr(_check_cons_edge, "py_func", _check_cons_edge)
-
-_resolve_incons_node = interpretation.resolve_inconsistency_node
-resolve_inconsistency_node = getattr(_resolve_incons_node, "py_func", _resolve_incons_node)
-
-_resolve_incons_edge = interpretation.resolve_inconsistency_edge
-resolve_inconsistency_edge = getattr(_resolve_incons_edge, "py_func", _resolve_incons_edge)
-
-_add_node_interp = interpretation._add_node_to_interpretation
-add_node_to_interpretation = getattr(_add_node_interp, "py_func", _add_node_interp)
-
-_add_edge_interp = interpretation._add_edge_to_interpretation
-add_edge_to_interpretation = getattr(_add_edge_interp, "py_func", _add_edge_interp)
-
-_add_edges_fn = interpretation._add_edges
-add_edges = getattr(_add_edges_fn, "py_func", _add_edges_fn)
-
-_delete_edge_fn = interpretation._delete_edge
-delete_edge = getattr(_delete_edge_fn, "py_func", _delete_edge_fn)
-
-_delete_node_fn = interpretation._delete_node
-delete_node = getattr(_delete_node_fn, "py_func", _delete_node_fn)
-
-_float_to_str = interpretation.float_to_str
-float_to_str = getattr(_float_to_str, "py_func", _float_to_str)
-
-_str_to_float = interpretation.str_to_float
-str_to_float = getattr(_str_to_float, "py_func", _str_to_float)
-
-_str_to_int = interpretation.str_to_int
-str_to_int = getattr(_str_to_int, "py_func", _str_to_int)
-
-_annotate = interpretation.annotate
-annotate = getattr(_annotate, "py_func", _annotate)
-
-_reason = interpretation.Interpretation.reason
-reason = getattr(_reason, "py_func", _reason)
+from tests.unit.disable_jit.interpretation_helpers import *
 
 
 class FakeWorld:
@@ -127,13 +23,12 @@ def interpretations():
         ('Justin', 'Dog'):  FakeWorld({'owns': True},  "owns,[1.0,1.0]"),
     }
 
-
 # ---- is_satisfied_node  and is_satisfied_edge tests ----
 
 def test_satisfied_path_true(interpretations):
     comp = ('Justin', 'Dog')
     na = ('owns', [1.0, 1.0])
-    assert _is_sat_node(interpretations, comp, na) is True
+    assert is_satisfied_node(interpretations, comp, na) is True
     assert is_satisfied_edge(interpretations, comp, na) is True
 
 def test_satisfied_path_false(interpretations):
@@ -146,7 +41,7 @@ def test_missing_comp_key_false():
     interpretations = {}
     comp = ('Nobody', 'Home')
     na = ('owns', None)
-    assert _is_sat_node(interpretations, comp, na) is True
+    assert is_satisfied_node(interpretations, comp, na) is True
     assert is_satisfied_edge(interpretations, comp, na) is True
     
 def test_is_satisfied_edge_returns_false_when_comp_missing():
@@ -154,7 +49,7 @@ def test_is_satisfied_edge_returns_false_when_comp_missing():
     interpretations = {}
     comp = ("ghost", "edge")
     na = ("owns", [1.0, 1.0])  # both non-None => enter try/except
-    assert _is_sat_node(interpretations, comp, na) is False
+    assert is_satisfied_node(interpretations, comp, na) is False
     assert is_satisfied_edge(interpretations, comp, na) is False
 
 
@@ -166,10 +61,10 @@ def test_get_qualified_edge_and_node_groundings_filters_true_edges(interpretatio
 
     # Separate mocks so each gets exactly 3 calls
     mock_is_sat_edge = Mock(side_effect=[False, True, True])  # F, T, T
-    mock_is_sat_node = Mock(side_effect=[False, True, True])  # F, T, T
+    mockis_satisfied_node = Mock(side_effect=[False, True, True])  # F, T, T
 
     monkeypatch.setattr(interpretation, "is_satisfied_edge", mock_is_sat_edge)
-    monkeypatch.setattr(interpretation, "is_satisfied_node", mock_is_sat_node)
+    monkeypatch.setattr(interpretation, "is_satisfied_node", mockis_satisfied_node)
 
     grounding = [
         ('Justin', 'Cat'),   # False
@@ -186,7 +81,7 @@ def test_get_qualified_edge_and_node_groundings_filters_true_edges(interpretatio
 
     # Each mock is called 3 times with the same argument sequence
     assert mock_is_sat_edge.call_count == 3
-    assert mock_is_sat_node.call_count == 3
+    assert mockis_satisfied_node.call_count == 3
 
     from unittest.mock import call
     expected_calls = [
@@ -195,7 +90,7 @@ def test_get_qualified_edge_and_node_groundings_filters_true_edges(interpretatio
         call(interpretations, grounding[2], (clause_l, clause_bnd)),
     ]
     mock_is_sat_edge.assert_has_calls(expected_calls)
-    mock_is_sat_node.assert_has_calls(expected_calls)
+    mockis_satisfied_node.assert_has_calls(expected_calls)
 
 
 def test_get_qualified_edge_and_node_groundings_none_qualify(interpretations, monkeypatch):
@@ -204,9 +99,9 @@ def test_get_qualified_edge_and_node_groundings_none_qualify(interpretations, mo
 
     # Separate mocks so each gets exactly len(grounding) calls
     mock_is_sat_edge = Mock(return_value=False)
-    mock_is_sat_node = Mock(return_value=False)
+    mockis_satisfied_node = Mock(return_value=False)
     monkeypatch.setattr(interpretation, "is_satisfied_edge", mock_is_sat_edge)
-    monkeypatch.setattr(interpretation, "is_satisfied_node", mock_is_sat_node)
+    monkeypatch.setattr(interpretation, "is_satisfied_node", mockis_satisfied_node)
 
     grounding = [('Justin', 'Dog'), ('Justin', 'Cat')]
 
@@ -216,23 +111,23 @@ def test_get_qualified_edge_and_node_groundings_none_qualify(interpretations, mo
     assert result_edge == []
     assert result_node == []
     assert mock_is_sat_edge.call_count == 2
-    assert mock_is_sat_node.call_count == 2
+    assert mockis_satisfied_node.call_count == 2
 
     expected_calls = [
         call(interpretations, grounding[0], ('owns', [1.0, 1.0])),
         call(interpretations, grounding[1], ('owns', [1.0, 1.0])),
     ]
     mock_is_sat_edge.assert_has_calls(expected_calls)
-    mock_is_sat_node.assert_has_calls(expected_calls)
+    mockis_satisfied_node.assert_has_calls(expected_calls)
 
 
 def test_get_qualified_edge_and_node_groundings_all_qualify(interpretations, monkeypatch):
     monkeypatch.setattr(interpretation.numba.typed.List, "empty_list", lambda *a, **k: [])
 
     mock_is_sat_edge = Mock(return_value=True)
-    mock_is_sat_node = Mock(return_value=True)
+    mockis_satisfied_node = Mock(return_value=True)
     monkeypatch.setattr(interpretation, "is_satisfied_edge", mock_is_sat_edge)
-    monkeypatch.setattr(interpretation, "is_satisfied_node", mock_is_sat_node)
+    monkeypatch.setattr(interpretation, "is_satisfied_node", mockis_satisfied_node)
 
     grounding = [('A', 'B'), ('C', 'D')]
 
@@ -242,14 +137,14 @@ def test_get_qualified_edge_and_node_groundings_all_qualify(interpretations, mon
     assert result_edge == grounding
     assert result_node == grounding
     assert mock_is_sat_edge.call_count == 2
-    assert mock_is_sat_node.call_count == 2
+    assert mockis_satisfied_node.call_count == 2
 
     expected_calls = [
         call(interpretations, grounding[0], ('owns', [1.0, 1.0])),
         call(interpretations, grounding[1], ('owns', [1.0, 1.0])),
     ]
     mock_is_sat_edge.assert_has_calls(expected_calls)
-    mock_is_sat_node.assert_has_calls(expected_calls)
+    mockis_satisfied_node.assert_has_calls(expected_calls)
 
 # ---- get_rule_node_clause_grounding tests ----
 
@@ -863,16 +758,6 @@ def test_add_node_minimal(monkeypatch):
     assert isinstance(interpretations_node["A"], DummyWorld)
 
 # ---- _add_edge tests (with _add_node mocked) ----
-
-class FakeLabel:
-    def __init__(self, value: str):
-        self.value = value
-    def __hash__(self):
-        return hash(self.value)
-    def __eq__(self, other):
-        return isinstance(other, FakeLabel) and self.value == other.value
-    def __repr__(self):
-        return f"FakeLabel({self.value!r})"
 
 def _shim_typed_list(monkeypatch):
     # Make typed.List(...) -> list(iterable) and .empty_list(...) -> []
@@ -1928,1233 +1813,6 @@ def test_ground_rule_node_recheck_failure_skips_body(monkeypatch):
     mock_check_all.assert_called_once()
     mock_add_node.assert_not_called()
     mock_add_edge.assert_not_called()
-
-
-# ---- reason function tests ----
-
-
-@pytest.fixture
-def reason_env(monkeypatch):
-    """Minimal environment to exercise Interpretation.reason."""
-
-    # Provide lightweight stand-ins for numba typed containers
-    class _ListShim:
-        def __call__(self, iterable=()):
-            return list(iterable)
-
-        def empty_list(self, *args, **kwargs):
-            return []
-
-    class _DictShim:
-        def empty(self, *args, **kwargs):
-            return {}
-
-    monkeypatch.setattr(interpretation.numba.typed, "List", _ListShim())
-    monkeypatch.setattr(interpretation.numba.typed, "Dict", _DictShim())
-    monkeypatch.setattr(interpretation.numba.types, "uint16", lambda x: x)
-
-    # Simple world/interval representations
-    class SimpleWorld:
-        def __init__(self):
-            self.world = {}
-
-    class SimpleInterval:
-        def __init__(self, val=1.0, static=False):
-            self.val = val
-            self._static = static
-
-        def copy(self):
-            return SimpleInterval(self.val, self._static)
-
-        def is_static(self):
-            return self._static
-
-    node = "n1"
-    lbl = label.Label("L")
-    bnd = SimpleInterval()
-
-    env = {
-        "interpretations_node": {0: {node: SimpleWorld()}},
-        "interpretations_edge": {0: {}},
-        "predicate_map_node": {},
-        "predicate_map_edge": {},
-        "tmax": 0,
-        "prev_reasoning_data": [0, 0],
-        "rules": [],
-        "nodes": [node],
-        "edges": [],
-        "neighbors": {node: []},
-        "reverse_neighbors": {node: []},
-        "rules_to_be_applied_node": [],
-        "rules_to_be_applied_edge": [],
-        "edges_to_be_added_node_rule": [],
-        "edges_to_be_added_edge_rule": [],
-        "rules_to_be_applied_node_trace": [],
-        "rules_to_be_applied_edge_trace": [],
-        "facts_to_be_applied_node": [(0, node, lbl, bnd, False, False)],
-        "facts_to_be_applied_edge": [],
-        "facts_to_be_applied_node_trace": [],
-        "facts_to_be_applied_edge_trace": [],
-        "ipl": [],
-        "rule_trace_node": [],
-        "rule_trace_edge": [],
-        "rule_trace_node_atoms": [],
-        "rule_trace_edge_atoms": [],
-        "reverse_graph": {},
-        "atom_trace": False,
-        "save_graph_attributes_to_rule_trace": False,
-        "persistent": False,
-        "inconsistency_check": False,
-        "store_interpretation_changes": False,
-        "update_mode": "",
-        "allow_ground_rules": True,
-        "max_facts_time": 0,
-        "annotation_functions": {},
-        "convergence_mode": "perfect_convergence",
-        "convergence_delta": 0,
-        "verbose": False,
-        "again": False,
-    }
-
-    def run(**overrides):
-        params = env.copy()
-        params.update(overrides)
-        return reason(
-            params["interpretations_node"],
-            params["interpretations_edge"],
-            params["predicate_map_node"],
-            params["predicate_map_edge"],
-            params["tmax"],
-            params["prev_reasoning_data"],
-            params["rules"],
-            params["nodes"],
-            params["edges"],
-            params["neighbors"],
-            params["reverse_neighbors"],
-            params["rules_to_be_applied_node"],
-            params["rules_to_be_applied_edge"],
-            params["edges_to_be_added_node_rule"],
-            params["edges_to_be_added_edge_rule"],
-            params["rules_to_be_applied_node_trace"],
-            params["rules_to_be_applied_edge_trace"],
-            params["facts_to_be_applied_node"],
-            params["facts_to_be_applied_edge"],
-            params["facts_to_be_applied_node_trace"],
-            params["facts_to_be_applied_edge_trace"],
-            params["ipl"],
-            params["rule_trace_node"],
-            params["rule_trace_edge"],
-            params["rule_trace_node_atoms"],
-            params["rule_trace_edge_atoms"],
-            params["reverse_graph"],
-            params["atom_trace"],
-            params["save_graph_attributes_to_rule_trace"],
-            params["persistent"],
-            params["inconsistency_check"],
-            params["store_interpretation_changes"],
-            params["update_mode"],
-            params["allow_ground_rules"],
-            params["max_facts_time"],
-            params["annotation_functions"],
-            params["convergence_mode"],
-            params["convergence_delta"],
-            params["verbose"],
-            params["again"],
-        )
-
-    env["run"] = run
-    env["node"] = node
-    env["label"] = lbl
-    env["bnd"] = bnd
-    return env
-
-
-def test_reason_applies_node_fact(monkeypatch, reason_env):
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: True)
-
-    def _updater(interp, predicate_map, comp, lb, *a, **k):
-        l, b = lb
-        interp[comp].world[l] = b
-        return True, 1
-
-    mock_update = Mock(side_effect=_updater)
-    monkeypatch.setattr(interpretation, "_update_node", mock_update)
-
-    fp, max_t = reason_env["run"]()
-
-    assert fp == 1 and max_t == 1
-    mock_update.assert_called_once()
-    assert mock_update.call_args.kwargs.get("override", False) is False
-    assert reason_env["interpretations_node"][0][reason_env["node"]].world[reason_env["label"]] is reason_env["bnd"]
-
-
-def test_reason_resolves_inconsistency(monkeypatch, reason_env):
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: False)
-    mock_resolve = Mock()
-    monkeypatch.setattr(interpretation, "resolve_inconsistency_node", mock_resolve)
-    monkeypatch.setattr(interpretation, "_update_node", Mock(return_value=(True, 1)))
-
-    reason_env["run"](inconsistency_check=True)
-
-    mock_resolve.assert_called_once()
-    interpretation._update_node.assert_not_called()
-
-
-def test_reason_overrides_without_inconsistency(monkeypatch, reason_env):
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: False)
-    mock_update = Mock(return_value=(True, 0))
-    monkeypatch.setattr(interpretation, "_update_node", mock_update)
-
-    reason_env["run"](inconsistency_check=False)
-
-    assert mock_update.call_args.kwargs.get("override") is True
-
-
-@pytest.mark.parametrize(
-    "mode,delta,change,expected_fp",
-    [
-        ("delta_interpretation", 1, 1, 0),
-        ("delta_interpretation", 0, 1, 1),
-        ("delta_bound", 0.5, 0.5, 0),
-        ("delta_bound", 0, 0.5, 1),
-    ],
-)
-def test_reason_convergence_modes(monkeypatch, reason_env, mode, delta, change, expected_fp):
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: True)
-
-    def _updater(interp, predicate_map, comp, lb, *a, **k):
-        l, b = lb
-        interp[comp].world[l] = b
-        return True, change
-
-    monkeypatch.setattr(interpretation, "_update_node", _updater)
-
-    fp, max_t = reason_env["run"](convergence_mode=mode, convergence_delta=delta)
-
-    assert fp == expected_fp and max_t == 1
-
-
-def make_copy_env(monkeypatch, persistent):
-    """Build a minimal environment starting from t=1 to test persistence copying."""
-
-    class _ListShim:
-        def __call__(self, iterable=()):
-            return list(iterable)
-
-        def empty_list(self, *a, **k):
-            return []
-
-    class _DictShim:
-        def empty(self, *a, **k):
-            return {}
-
-    monkeypatch.setattr(interpretation.numba.typed, "List", _ListShim())
-    monkeypatch.setattr(interpretation.numba.typed, "Dict", _DictShim())
-    monkeypatch.setattr(interpretation.numba.types, "uint16", lambda x: x)
-
-    class SimpleWorld:
-        def __init__(self, *a, **k):
-            self.world = {}
-
-    monkeypatch.setattr(interpretation.world, "World", SimpleWorld)
-
-    class SimpleInterval:
-        def __init__(self, static=False):
-            self._static = static
-
-        def copy(self):
-            return SimpleInterval(self._static)
-
-        def is_static(self):
-            return self._static
-
-    n1, n2 = "n1", "n2"
-    edge = (n1, n2)
-    dyn_lbl = label.Label("dyn")
-    stat_lbl = label.Label("stat")
-
-    node_w0 = SimpleWorld()
-    node_w0.world[dyn_lbl] = SimpleInterval()
-    node_w0.world[stat_lbl] = SimpleInterval(static=True)
-
-    edge_w0 = SimpleWorld()
-    edge_w0.world[dyn_lbl] = SimpleInterval()
-    edge_w0.world[stat_lbl] = SimpleInterval(static=True)
-
-    env = {
-        "interpretations_node": {0: {n1: node_w0}},
-        "interpretations_edge": {0: {edge: edge_w0}},
-        "predicate_map_node": {},
-        "predicate_map_edge": {},
-        "tmax": 1,
-        "prev_reasoning_data": [1, 0],
-        "rules": [],
-        "nodes": [n1, n2],
-        "edges": [edge],
-        "neighbors": {n1: [n2], n2: []},
-        "reverse_neighbors": {n1: [], n2: [n1]},
-        "rules_to_be_applied_node": [],
-        "rules_to_be_applied_edge": [],
-        "edges_to_be_added_node_rule": [],
-        "edges_to_be_added_edge_rule": [],
-        "rules_to_be_applied_node_trace": [],
-        "rules_to_be_applied_edge_trace": [],
-        "facts_to_be_applied_node": [],
-        "facts_to_be_applied_edge": [],
-        "facts_to_be_applied_node_trace": [],
-        "facts_to_be_applied_edge_trace": [],
-        "ipl": [],
-        "rule_trace_node": [],
-        "rule_trace_edge": [],
-        "rule_trace_node_atoms": [],
-        "rule_trace_edge_atoms": [],
-        "reverse_graph": {},
-        "atom_trace": False,
-        "save_graph_attributes_to_rule_trace": False,
-        "persistent": persistent,
-        "inconsistency_check": False,
-        "store_interpretation_changes": False,
-        "update_mode": "",
-        "allow_ground_rules": True,
-        "max_facts_time": 0,
-        "annotation_functions": {},
-        "convergence_mode": "perfect_convergence",
-        "convergence_delta": 0,
-        "verbose": False,
-        "again": False,
-    }
-
-    def run(**overrides):
-        params = env.copy()
-        params.update(overrides)
-        return reason(
-            params["interpretations_node"],
-            params["interpretations_edge"],
-            params["predicate_map_node"],
-            params["predicate_map_edge"],
-            params["tmax"],
-            params["prev_reasoning_data"],
-            params["rules"],
-            params["nodes"],
-            params["edges"],
-            params["neighbors"],
-            params["reverse_neighbors"],
-            params["rules_to_be_applied_node"],
-            params["rules_to_be_applied_edge"],
-            params["edges_to_be_added_node_rule"],
-            params["edges_to_be_added_edge_rule"],
-            params["rules_to_be_applied_node_trace"],
-            params["rules_to_be_applied_edge_trace"],
-            params["facts_to_be_applied_node"],
-            params["facts_to_be_applied_edge"],
-            params["facts_to_be_applied_node_trace"],
-            params["facts_to_be_applied_edge_trace"],
-            params["ipl"],
-            params["rule_trace_node"],
-            params["rule_trace_edge"],
-            params["rule_trace_node_atoms"],
-            params["rule_trace_edge_atoms"],
-            params["reverse_graph"],
-            params["atom_trace"],
-            params["save_graph_attributes_to_rule_trace"],
-            params["persistent"],
-            params["inconsistency_check"],
-            params["store_interpretation_changes"],
-            params["update_mode"],
-            params["allow_ground_rules"],
-            params["max_facts_time"],
-            params["annotation_functions"],
-            params["convergence_mode"],
-            params["convergence_delta"],
-            params["verbose"],
-            params["again"],
-        )
-
-    env["run"] = run
-    env["node"] = n1
-    env["edge"] = edge
-    env["dyn_label"] = dyn_lbl
-    env["stat_label"] = stat_lbl
-    return env
-
-
-def test_reason_breaks_when_no_update(monkeypatch, reason_env):
-    fp, max_t = reason_env["run"](
-        facts_to_be_applied_node=[],
-        convergence_mode="delta_bound",
-        convergence_delta=-1,
-    )
-    assert fp == 1 and max_t == 1
-
-
-@pytest.mark.parametrize("persistent", [True, False])
-def test_reason_copies_previous_timestep(monkeypatch, persistent):
-    env = make_copy_env(monkeypatch, persistent)
-    env["run"]()
-
-    expected = {env["stat_label"]} | ({env["dyn_label"]} if persistent else set())
-
-    node_world = env["interpretations_node"][1][env["node"]].world
-    edge_world = env["interpretations_edge"][1][env["edge"]].world
-
-    assert set(node_world.keys()) == expected
-    assert set(edge_world.keys()) == expected
-
-
-def test_reason_adds_missing_node(monkeypatch, reason_env):
-    new_node = "n2"
-
-    def stub_add_node(node, neighbors, reverse_neighbors, nodes, interp):
-        nodes.append(node)
-        neighbors[node] = []
-        reverse_neighbors[node] = []
-        interp[node] = reason_env["interpretations_node"][0][reason_env["node"]].__class__()
-
-    monkeypatch.setattr(interpretation, "_add_node", stub_add_node)
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: True)
-    monkeypatch.setattr(interpretation, "_update_node", lambda *a, **k: (False, 0))
-
-    nodes = [reason_env["node"]]
-    neighbors = {reason_env["node"]: []}
-    reverse = {reason_env["node"]: []}
-    interp = {0: {reason_env["node"]: reason_env["interpretations_node"][0][reason_env["node"]]}}
-    facts = [(0, new_node, reason_env["label"], reason_env["bnd"], False, False)]
-
-    reason_env["run"](
-        nodes=nodes,
-        neighbors=neighbors,
-        reverse_neighbors=reverse,
-        interpretations_node=interp,
-        facts_to_be_applied_node=facts,
-    )
-
-    assert new_node in nodes and new_node in interp[0] and neighbors[new_node] == []
-
-
-def test_reason_logs_static_fact(monkeypatch, reason_env):
-    node = reason_env["node"]
-    label_ = reason_env["label"]
-    static_bnd = reason_env["bnd"].__class__(1.0, True)
-    reason_env["interpretations_node"][0][node].world[label_] = static_bnd
-    new_bnd = reason_env["bnd"].__class__(0.5, False)
-    facts = [(0, node, label_, new_bnd, False, False)]
-    rule_trace = []
-
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: True)
-
-    reason_env["run"](
-        facts_to_be_applied_node=facts,
-        store_interpretation_changes=True,
-        rule_trace_node=rule_trace,
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert rule_trace == [(0, 1, node, label_, new_bnd)]
-    assert reason_env["interpretations_node"][0][node].world[label_] is static_bnd
-
-
-def test_reason_static_fact_traces_and_requeues(reason_env):
-    node = reason_env["node"]
-    lbl = reason_env["label"]
-    other = label.Label("other")
-    static_bnd = reason_env["bnd"].__class__(1.0, True)
-    other_bnd = reason_env["bnd"].__class__(0.7, False)
-    reason_env["interpretations_node"][0][node].world[lbl] = static_bnd
-    reason_env["interpretations_node"][0][node].world[other] = other_bnd
-    new_bnd = reason_env["bnd"].__class__(0.2, False)
-    facts = [(0, node, lbl, new_bnd, True, False)]
-    trace = [["x"]]
-    rule_trace = []
-    rule_trace_atoms = []
-    ipl = [(lbl, other)]
-
-    reason_env["run"](
-        facts_to_be_applied_node=facts,
-        facts_to_be_applied_node_trace=trace,
-        rule_trace_node=rule_trace,
-        rule_trace_node_atoms=rule_trace_atoms,
-        atom_trace=True,
-        store_interpretation_changes=True,
-        ipl=ipl,
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert facts == [(1, node, lbl, new_bnd, True, False)]
-    assert trace == [["x"]]
-    assert rule_trace == [(0, 1, node, lbl, new_bnd), (0, 1, node, other, other_bnd)]
-    assert len(rule_trace_atoms) == 2
-    assert reason_env["interpretations_node"][0][node].world[lbl] is static_bnd
-
-
-def test_reason_static_fact_traces_complement_second(reason_env):
-    node = reason_env["node"]
-    lbl = reason_env["label"]
-    other = label.Label("other")
-    static_bnd = reason_env["bnd"].__class__(1.0, True)
-    other_bnd = reason_env["bnd"].__class__(0.3, False)
-    reason_env["interpretations_node"][0][node].world[lbl] = static_bnd
-    reason_env["interpretations_node"][0][node].world[other] = other_bnd
-    new_bnd = reason_env["bnd"].__class__(0.4, False)
-    facts = [(0, node, lbl, new_bnd, True, False)]
-    trace = [["z"]]
-    rule_trace = []
-    rule_trace_atoms = []
-    ipl = [(other, lbl)]
-
-    reason_env["run"](
-        facts_to_be_applied_node=facts,
-        facts_to_be_applied_node_trace=trace,
-        rule_trace_node=rule_trace,
-        rule_trace_node_atoms=rule_trace_atoms,
-        atom_trace=True,
-        store_interpretation_changes=True,
-        ipl=ipl,
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert facts == [(1, node, lbl, new_bnd, True, False)]
-    assert trace == [["z"]]
-    assert rule_trace == [(0, 1, node, lbl, new_bnd), (0, 1, node, other, other_bnd)]
-    assert len(rule_trace_atoms) == 2
-    assert reason_env["interpretations_node"][0][node].world[lbl] is static_bnd
-
-
-def test_reason_delta_bound_inconsistent(monkeypatch, reason_env):
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: False)
-
-    def _updater(interp, predicate_map, comp, lb, *a, **k):
-        l, b = lb
-        interp[comp].world[l] = b
-        return True, 0.5
-
-    monkeypatch.setattr(interpretation, "_update_node", _updater)
-
-    fp, _ = reason_env["run"](
-        convergence_mode="delta_bound",
-        convergence_delta=0,
-    )
-
-    assert fp == 1
-
-
-def test_reason_defers_future_fact_and_traces(reason_env):
-    future_fact = [(1, reason_env["node"], reason_env["label"], reason_env["bnd"], False, False)]
-    future_trace = [["t"]]
-
-    reason_env["run"](
-        facts_to_be_applied_node=future_fact,
-        facts_to_be_applied_node_trace=future_trace,
-        atom_trace=True,
-    )
-
-    assert future_fact == [(1, reason_env["node"], reason_env["label"], reason_env["bnd"], False, False)]
-    assert future_trace == [["t"]]
-
-
-def test_reason_defers_future_edge_fact_and_traces(reason_env):
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    bnd = reason_env["bnd"]
-    edge_world = reason_env["interpretations_node"][0][node].__class__()
-    interpretations_edge = {0: {edge: edge_world}}
-    edges = [edge]
-    future_fact = [(1, edge, lbl, bnd, False, False)]
-    future_trace = ["edge"]
-
-    reason_env["run"](
-        edges=edges,
-        neighbors={node: []},
-        reverse_neighbors={node: []},
-        interpretations_edge=interpretations_edge,
-        facts_to_be_applied_edge=future_fact,
-        facts_to_be_applied_edge_trace=future_trace,
-        facts_to_be_applied_node=[],
-        atom_trace=True,
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert future_fact == [(1, edge, lbl, bnd, False, False)]
-    assert future_trace == ["edge"]
-
-
-def test_reason_applies_edge_fact(monkeypatch, reason_env):
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    bnd = reason_env["bnd"]
-    edge_world = reason_env["interpretations_node"][0][node].__class__()
-    interpretations_edge = {0: {edge: edge_world}}
-    edges = [edge]
-    neighbors = {node: []}
-    reverse = {node: []}
-    facts = [(0, edge, lbl, bnd, False, False)]
-
-    monkeypatch.setattr(interpretation, "check_consistent_edge", lambda *a, **k: True)
-
-    called = {}
-
-    def _update_edge_stub(interp, predicate_map, comp, lb, *a, **k):
-        l, bound = lb
-        interp[comp].world[l] = bound
-        called["ok"] = True
-        return True, 0
-
-    monkeypatch.setattr(interpretation, "_update_edge", _update_edge_stub)
-
-    reason_env["run"](
-        edges=edges,
-        neighbors=neighbors,
-        reverse_neighbors=reverse,
-        interpretations_edge=interpretations_edge,
-        facts_to_be_applied_edge=facts,
-        facts_to_be_applied_node=[],
-        predicate_map_edge={},
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert called.get("ok")
-    assert edge_world.world[lbl] is bnd
-
-
-def test_reason_adds_missing_edge(monkeypatch, reason_env):
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    bnd = reason_env["bnd"]
-    edges = []
-    neighbors = {node: []}
-    reverse = {node: []}
-    interpretations_edge = {0: {}}
-    called = {}
-
-    def fake_add_edge(s, t, nbrs, rev, nodes, edges_list, l, interp_node, interp_edge, pred_map, t_val):
-        interp_edge[(s, t)] = reason_env["interpretations_node"][0][node].__class__()
-        edges_list.append((s, t))
-        called["ok"] = True
-        return (s, t), True
-
-    monkeypatch.setattr(interpretation, "_add_edge", fake_add_edge)
-    monkeypatch.setattr(interpretation, "check_consistent_edge", lambda *a, **k: True)
-
-    def updater(interp, predicate_map, comp, lb, *a, **k):
-        l, bound = lb
-        interp[comp].world[l] = bound
-        return True, 0
-
-    monkeypatch.setattr(interpretation, "_update_edge", updater)
-
-    reason_env["run"](
-        edges=edges,
-        neighbors=neighbors,
-        reverse_neighbors=reverse,
-        interpretations_edge=interpretations_edge,
-        facts_to_be_applied_edge=[(0, edge, lbl, bnd, False, False)],
-        facts_to_be_applied_node=[],
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert called.get("ok")
-    assert edge in edges
-    assert edge in interpretations_edge[0]
-
-
-def test_reason_adds_edge_to_interpretation(monkeypatch, reason_env):
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    bnd = reason_env["bnd"]
-    edges = [edge]
-    neighbors = {node: []}
-    reverse = {node: []}
-    interpretations_edge = {0: {}}
-    called = {}
-
-    def fake_add_edge_to_interp(edge_comp, interp_edge):
-        interp_edge[edge_comp] = reason_env["interpretations_node"][0][node].__class__()
-        called["ok"] = True
-
-    monkeypatch.setattr(interpretation, "_add_edge_to_interpretation", fake_add_edge_to_interp)
-    monkeypatch.setattr(interpretation, "check_consistent_edge", lambda *a, **k: True)
-
-    def updater(interp, predicate_map, comp, lb, *a, **k):
-        l, bound = lb
-        interp[comp].world[l] = bound
-        return True, 0
-
-    monkeypatch.setattr(interpretation, "_update_edge", updater)
-
-    reason_env["run"](
-        edges=edges,
-        neighbors=neighbors,
-        reverse_neighbors=reverse,
-        interpretations_edge=interpretations_edge,
-        facts_to_be_applied_edge=[(0, edge, lbl, bnd, False, False)],
-        facts_to_be_applied_node=[],
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert called.get("ok")
-    assert edge in interpretations_edge[0]
-
-
-@pytest.mark.parametrize(
-    "save_attrs, graph_attr, store_changes, expect_trace",
-    [
-        (True, True, True, True),
-        (False, True, True, False),
-        (True, True, False, False),
-        (False, False, True, True),
-    ],
-)
-def test_reason_static_edge_rule_trace_branches(
-    reason_env, save_attrs, graph_attr, store_changes, expect_trace
-):
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    other = label.Label("other")
-    static_bnd = reason_env["bnd"].__class__(1.0, True)
-    other_bnd = reason_env["bnd"].__class__(0.5, False)
-    world = reason_env["interpretations_node"][0][node].__class__()
-    world.world = {lbl: static_bnd, other: other_bnd}
-    interpretations_edge = {0: {edge: world}}
-    edges = [edge]
-    facts = [(0, edge, lbl, reason_env["bnd"], True, graph_attr)]
-    rule_trace = []
-
-    reason_env["run"](
-        edges=edges,
-        neighbors={node: []},
-        reverse_neighbors={node: []},
-        interpretations_edge=interpretations_edge,
-        facts_to_be_applied_edge=facts,
-        facts_to_be_applied_node=[],
-        rule_trace_edge=rule_trace,
-        store_interpretation_changes=store_changes,
-        save_graph_attributes_to_rule_trace=save_attrs,
-        ipl=[(lbl, other)],
-        prev_reasoning_data=[0, 1],
-    )
-
-    if expect_trace:
-        assert rule_trace == [
-            (0, 1, edge, lbl, static_bnd),
-            (0, 1, edge, other, other_bnd),
-        ]
-    else:
-        assert rule_trace == []
-    assert facts == [(1, edge, lbl, reason_env["bnd"], True, graph_attr)]
-
-
-def test_reason_static_edge_atom_trace_complements(monkeypatch, reason_env):
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    other1 = label.Label("o1")
-    other2 = label.Label("o2")
-    static_bnd = reason_env["bnd"].__class__(1.0, True)
-    o1_bnd = reason_env["bnd"].__class__(0.5, False)
-    o2_bnd = reason_env["bnd"].__class__(0.6, False)
-    world = reason_env["interpretations_node"][0][node].__class__()
-    world.world = {lbl: static_bnd, other1: o1_bnd, other2: o2_bnd}
-    interpretations_edge = {0: {edge: world}}
-    edges = [edge]
-    facts = [(0, edge, lbl, reason_env["bnd"], True, False)]
-    facts_trace = ["t"]
-    rule_trace = []
-    rule_trace_atoms = []
-    mock_update = Mock()
-    monkeypatch.setattr(interpretation, "_update_rule_trace", mock_update)
-
-    reason_env["run"](
-        edges=edges,
-        neighbors={node: []},
-        reverse_neighbors={node: []},
-        interpretations_edge=interpretations_edge,
-        facts_to_be_applied_edge=facts,
-        facts_to_be_applied_edge_trace=facts_trace,
-        facts_to_be_applied_node=[],
-        rule_trace_edge=rule_trace,
-        rule_trace_edge_atoms=rule_trace_atoms,
-        store_interpretation_changes=True,
-        atom_trace=True,
-        ipl=[(lbl, other1), (other2, lbl)],
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert rule_trace == [
-        (0, 1, edge, lbl, static_bnd),
-        (0, 1, edge, other1, o1_bnd),
-        (0, 1, edge, other2, o2_bnd),
-    ]
-    assert facts == [(1, edge, lbl, reason_env["bnd"], True, False)]
-    assert facts_trace == ["t"]
-    assert mock_update.call_count == 3
-    calls = [
-        call(rule_trace_atoms, [], [], reason_env["bnd"], "t"),
-        call(rule_trace_atoms, [], [], o1_bnd, "t"),
-        call(rule_trace_atoms, [], [], o2_bnd, "t"),
-    ]
-    mock_update.assert_has_calls(calls)
-
-
-def test_reason_edge_delta_bound(monkeypatch, reason_env):
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    bnd = reason_env["bnd"]
-    world = reason_env["interpretations_node"][0][node].__class__()
-    interpretations_edge = {0: {edge: world}}
-    edges = [edge]
-
-    monkeypatch.setattr(interpretation, "check_consistent_edge", lambda *a, **k: True)
-
-    def _update_edge_stub(interp, predicate_map, comp, lb, *a, **k):
-        l, bound = lb
-        interp[comp].world[l] = bound
-        return True, 0.5
-
-    monkeypatch.setattr(interpretation, "_update_edge", _update_edge_stub)
-
-    fp, _ = reason_env["run"](
-        edges=edges,
-        neighbors={node: []},
-        reverse_neighbors={node: []},
-        interpretations_edge=interpretations_edge,
-        facts_to_be_applied_edge=[(0, edge, lbl, bnd, False, False)],
-        facts_to_be_applied_node=[],
-        convergence_mode="delta_bound",
-        convergence_delta=0,
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert fp == 2
-
-
-@pytest.mark.parametrize("inconsistency_check", [True, False])
-def test_reason_edge_inconsistency_branches(monkeypatch, reason_env, inconsistency_check):
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    bnd = reason_env["bnd"]
-    world = reason_env["interpretations_node"][0][node].__class__()
-    interpretations_edge = {0: {edge: world}}
-    edges = [edge]
-
-    monkeypatch.setattr(interpretation, "check_consistent_edge", lambda *a, **k: False)
-    mock_resolve = Mock()
-    mock_update = Mock(return_value=(True, 0))
-    monkeypatch.setattr(interpretation, "resolve_inconsistency_edge", mock_resolve)
-    monkeypatch.setattr(interpretation, "_update_edge", mock_update)
-
-    reason_env["run"](
-        edges=edges,
-        neighbors={node: []},
-        reverse_neighbors={node: []},
-        interpretations_edge=interpretations_edge,
-        facts_to_be_applied_edge=[(0, edge, lbl, bnd, False, False)],
-        facts_to_be_applied_node=[],
-        inconsistency_check=inconsistency_check,
-        prev_reasoning_data=[0, 1],
-    )
-
-    if inconsistency_check:
-        assert mock_resolve.called
-        assert not mock_update.called
-    else:
-        assert mock_update.called
-        assert mock_update.call_args[1]["override"] is True
-        assert not mock_resolve.called
-
-
-def test_reason_edge_inconsistency_delta_bound_override(monkeypatch, reason_env):
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: True)
-    monkeypatch.setattr(interpretation, "_update_node", lambda *a, **k: (True, 0))
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    bnd = reason_env["bnd"]
-    world = reason_env["interpretations_node"][0][node].__class__()
-    interpretations_edge = {0: {edge: world}}
-    edges = [edge]
-
-    monkeypatch.setattr(interpretation, "check_consistent_edge", lambda *a, **k: False)
-
-    def _update_edge_stub(interp, predicate_map, comp, lb, *a, **k):
-        l, bound = lb
-        interp[comp].world[l] = bound
-        return True, 0.5
-
-    monkeypatch.setattr(interpretation, "_update_edge", _update_edge_stub)
-
-    fp, _ = reason_env["run"](
-        edges=edges,
-        neighbors={node: []},
-        reverse_neighbors={node: []},
-        interpretations_edge=interpretations_edge,
-        facts_to_be_applied_edge=[(0, edge, lbl, bnd, False, False)],
-        facts_to_be_applied_node=[],
-        convergence_mode="delta_bound",
-        convergence_delta=0,
-        inconsistency_check=False,
-        prev_reasoning_data=[0, 1],
-    )
-
-    assert fp == 2
-
-
-def test_reason_skips_rule_outside_tmax(monkeypatch, reason_env):
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: True)
-    monkeypatch.setattr(interpretation, "_update_node", lambda *a, **k: (True, 0))
-    rule = Mock()
-    rule.get_delta.return_value = 1
-    rule.get_target.return_value = reason_env["label"]
-    rule.is_static_rule.return_value = False
-    rule.get_weights.return_value = []
-    rule.get_annotation_function.return_value = ""
-    rule.get_bnd.return_value = Mock(lower=0, upper=1)
-
-    mock_ground = Mock(return_value=([], []))
-    monkeypatch.setattr(interpretation, "_ground_rule", mock_ground)
-
-    fp, _ = reason_env["run"](rules=[rule], tmax=0)
-
-    assert fp == 1
-    mock_ground.assert_not_called()
-
-
-
-
-
-
-def test_reason_edge_rule_records_trace(monkeypatch, reason_env):
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: True)
-    monkeypatch.setattr(interpretation, "_update_node", lambda *a, **k: (True, 0))
-    monkeypatch.setattr(interpretation, "annotate", lambda *a, **k: (0, 1))
-    monkeypatch.setattr(interpretation.interval, "closed", lambda lo, up: reason_env["bnd"].__class__(lo, False))
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    interpretations_edge = {0: {}}
-    edges = [edge]
-
-    rule = Mock()
-    rule.get_delta.return_value = 1
-    rule.get_target.return_value = lbl
-    rule.is_static_rule.return_value = False
-    rule.get_name.return_value = "r"
-    rule.get_weights.return_value = []
-    rule.get_annotation_function.return_value = ""
-    rule.get_bnd.return_value = Mock(lower=0, upper=1)
-
-    edge_placeholder = type("EL", (), {"value": ""})()
-    applicable = [(edge, [], [node], [edge], ([], [], edge_placeholder))]
-    monkeypatch.setattr(interpretation, "_ground_rule", lambda *a, **k: ([], applicable))
-
-    def add_edge_to_interp(comp, interp_edge):
-        interp_edge[comp] = type("W", (), {"world": {}})()
-
-    monkeypatch.setattr(interpretation, "_add_edge_to_interpretation", add_edge_to_interp)
-    monkeypatch.setattr(interpretation, "check_consistent_edge", lambda *a, **k: True)
-
-    captured = {}
-
-    def update_edge_stub(*args, **kwargs):
-        rt_trace = args[12]
-        idx = args[13]
-        captured["trace"] = rt_trace[idx]
-        return False, 0
-
-    monkeypatch.setattr(interpretation, "_update_edge", update_edge_stub)
-
-    reason_env["run"](
-        rules=[rule],
-        edges=edges,
-        neighbors={node: []},
-        reverse_neighbors={node: []},
-        interpretations_edge=interpretations_edge,
-        atom_trace=True,
-        tmax=1,
-    )
-
-    assert captured["trace"] == ([node], [edge], "r")
-
-
-def test_reason_edge_rule_delta_zero_applies(monkeypatch, reason_env):
-    monkeypatch.setattr(interpretation, "check_consistent_node", lambda *a, **k: True)
-    monkeypatch.setattr(interpretation, "_update_node", lambda *a, **k: (True, 0))
-    monkeypatch.setattr(interpretation, "annotate", lambda *a, **k: (0, 1))
-    monkeypatch.setattr(interpretation.interval, "closed", lambda lo, up: reason_env["bnd"].__class__(lo, False))
-    node = reason_env["node"]
-    edge = (node, node)
-    lbl = reason_env["label"]
-    interpretations_edge = {0: {}}
-    edges = [edge]
-
-    rule = Mock()
-    rule.get_delta.return_value = 0
-    rule.get_target.return_value = lbl
-    rule.is_static_rule.return_value = False
-    rule.get_weights.return_value = []
-    rule.get_annotation_function.return_value = ""
-    rule.get_bnd.return_value = Mock(lower=0, upper=1)
-
-    edge_placeholder = type("EL", (), {"value": ""})()
-    applicable = [(edge, [], [], [], ([], [], edge_placeholder))]
-    monkeypatch.setattr(interpretation, "_ground_rule", lambda *a, **k: ([], applicable))
-
-    def add_edge_to_interp(comp, interp_edge):
-        interp_edge[comp] = type("W", (), {"world": {}})()
-
-    monkeypatch.setattr(interpretation, "_add_edge_to_interpretation", add_edge_to_interp)
-    monkeypatch.setattr(interpretation, "check_consistent_edge", lambda *a, **k: True)
-
-    called = {}
-
-    def update_edge_stub(*a, **k):
-        called["ok"] = True
-        return False, 0
-
-    monkeypatch.setattr(interpretation, "_update_edge", update_edge_stub)
-
-    reason_env["run"](
-        rules=[rule],
-        edges=edges,
-        neighbors={node: []},
-        reverse_neighbors={node: []},
-        interpretations_edge=interpretations_edge,
-        tmax=0,
-    )
-
-    assert called.get("ok")
-
-
-# ---- check_consistent_node / check_consistent_edge tests ----
-
-class _Interval:
-    def __init__(self, lower, upper):
-        self.lower = lower
-        self.upper = upper
-        self.static = False
-    def copy(self):
-        return _Interval(self.lower, self.upper)
-    def set_lower_upper(self, lo, up):
-        self.lower, self.upper = lo, up
-    def set_static(self, val):
-        self.static = val
-
-class _World:
-    def __init__(self, mapping=None):
-        self.world = mapping or {}
-    def is_satisfied(self, label, bnd):
-        w = self.world[label]
-        return not (bnd.lower > w.upper or w.lower > bnd.upper)
-
-@pytest.mark.parametrize("check_fn", [check_consistent_node, check_consistent_edge])
-def test_check_consistent_functions(monkeypatch, check_fn):
-    monkeypatch.setattr(interpretation.interval, "closed", lambda lo, up: _Interval(lo, up))
-    interp = {"c": _World({"p": _Interval(0, 0.5)})}
-    assert check_fn(interp, "c", ("p", _Interval(0.4, 0.6))) is True
-    assert check_fn(interp, "c", ("p", _Interval(0.6, 0.8))) is False
-    interp2 = {"c": _World({})}
-    assert check_fn(interp2, "c", ("p", _Interval(0.6, 0.8))) is True
-
-
-# ---- resolve_inconsistency_node / resolve_inconsistency_edge tests ----
-
-@pytest.mark.parametrize(
-    "resolver,comp_key",
-    [
-        (resolve_inconsistency_node, "n"),
-        (resolve_inconsistency_edge, ("s", "t")),
-    ],
-)
-def test_resolve_inconsistency_updates_world_and_trace(monkeypatch, resolver, comp_key):
-    monkeypatch.setattr(interpretation.interval, "closed", lambda lo, up: _Interval(lo, up))
-    calls = []
-    monkeypatch.setattr(interpretation, "_update_rule_trace", lambda *a: calls.append(a))
-    world = _World({"p": _Interval(0, 0.5), "q": _Interval(0, 0.5), "r": _Interval(0, 0.5)})
-    interpretations = {comp_key: world}
-    ipl = [("p", "q"), ("r", "p")]
-    rule_trace = []
-    rule_trace_atoms = []
-    facts = ["fact"]
-    resolver(
-        interpretations,
-        comp_key,
-        ("p", _Interval(0.9, 1.0)),
-        ipl,
-        1,
-        2,
-        0,
-        True,
-        rule_trace,
-        rule_trace_atoms,
-        [],
-        facts,
-        True,
-        "fact",
-    )
-    assert world.world["p"].lower == 0 and world.world["p"].upper == 1 and world.world["p"].static
-    assert world.world["q"].lower == 0 and world.world["q"].upper == 1 and world.world["q"].static
-    assert world.world["r"].lower == 0 and world.world["r"].upper == 1 and world.world["r"].static
-    assert len(rule_trace) == 3
-    assert len(calls) == 3
-
-
-# ---- _add_node_to_interpretation / _add_edge_to_interpretation tests ----
-
-def test_add_node_and_edge_to_interpretation(monkeypatch):
-    class DummyWorld:
-        def __init__(self, labels):
-            self.labels = labels
-    monkeypatch.setattr(interpretation.world, "World", DummyWorld)
-    nodes = {}
-    add_node_to_interpretation("A", nodes)
-    assert isinstance(nodes["A"], DummyWorld)
-    edges = {}
-    add_edge_to_interpretation(("A", "B"), edges)
-    assert isinstance(edges[("A", "B")], DummyWorld)
-
-
-# ---- _add_edges tests ----
-
-def test_add_edges_counts_new_edges(monkeypatch):
-    def fake_add_edge(src, tgt, neighbors, reverse_neighbors, nodes, edges, l, interp_node, interp_edge, pred, t):
-        edge = (src, tgt)
-        new_edge = edge not in edges
-        if new_edge:
-            edges.append(edge)
-        return edge, new_edge
-    monkeypatch.setattr(interpretation, "_add_edge", fake_add_edge)
-    edges = [("A", "B")]
-    added, changes = add_edges(["A"], ["B", "C"], {}, {}, [], edges, FakeLabel("L"), {}, {}, {}, 0)
-    assert added == [("A", "B"), ("A", "C")]
-    assert changes == 1
-
-
-# ---- _delete_edge / _delete_node tests ----
-
-def test_delete_edge_removes_all_references():
-    lbl = FakeLabel("L")
-    neighbors = {"A": ["B"], "C": []}
-    reverse_neighbors = {"B": ["A"], "C": []}
-    edges = [("A", "B")]
-    interp_edge = {("A", "B"): "W"}
-    predicate_map = {lbl: [("A", "B"), ("C", "D")]}
-    delete_edge(("A", "B"), neighbors, reverse_neighbors, edges, interp_edge, predicate_map)
-    assert edges == []
-    assert interp_edge == {}
-    assert neighbors["A"] == [] and reverse_neighbors["B"] == []
-    assert predicate_map[lbl] == [("C", "D")]
-
-
-def test_delete_node_removes_all_references():
-    lbl = FakeLabel("L")
-    neighbors = {"A": ["B"], "B": [], "C": ["A"]}
-    reverse_neighbors = {"A": ["C"], "B": ["A"], "C": []}
-    nodes = ["A", "B", "C"]
-    interp_node = {"A": "wA", "B": "wB", "C": "wC"}
-    predicate_map = {lbl: ["A", "B"]}
-    delete_node("A", neighbors, reverse_neighbors, nodes, interp_node, predicate_map)
-    assert nodes == ["B", "C"]
-    assert "A" not in neighbors and "A" not in reverse_neighbors
-    assert interp_node == {"B": "wB", "C": "wC"}
-    assert predicate_map[lbl] == ["B"]
-
-
-# ---- are_satisfied_node / are_satisfied_edge tests ----
-
-@pytest.mark.parametrize(
-    "are_fn,sat_name",
-    [
-        (are_satisfied_node, "is_satisfied_node"),
-        (are_satisfied_edge, "is_satisfied_edge"),
-    ],
-)
-def test_are_satisfied_helpers_call_each(monkeypatch, are_fn, sat_name):
-    mock = Mock(side_effect=[True, False])
-    monkeypatch.setattr(interpretation, sat_name, mock)
-    nas = [("l1", _Interval(0, 1)), ("l2", _Interval(0, 1))]
-    out = are_fn({}, "c", nas)
-    assert out is False
-    expected = [call({}, "c", nas[0]), call({}, "c", nas[1])]
-    mock.assert_has_calls(expected)
-
-
-# ---- is_satisfied_*_comparison tests ----
-
-@pytest.mark.parametrize(
-    "cmp_fn,interp_key",
-    [
-        (is_satisfied_node_comparison, "n"),
-        (is_satisfied_edge_comparison, ("s", "t")),
-    ],
-)
-def test_is_satisfied_comparison(monkeypatch, cmp_fn, interp_key):
-    monkeypatch.setattr(interpretation, "str_to_float", lambda s: float(s))
-    w = _World({FakeLabel("p.5"): _Interval(0, 1)})
-    interpretations = {interp_key: w}
-    res, num = cmp_fn(interpretations, interp_key, (FakeLabel("p"), _Interval(0, 1)))
-    assert res is True and num == 5.0
-    res, num = cmp_fn(interpretations, interp_key, (FakeLabel("q"), _Interval(0, 1)))
-    assert res is False and num == 0
-
-
-# ---- _update_rule_trace tests ----
-
-def test_update_rule_trace_makes_copy():
-    rt = []
-    bnd = _Interval(0.1, 0.2)
-    update_rule_trace(rt, [["n1"]], [[("a", "b")]], bnd, "name")
-    assert rt[0][0] == [["n1"]]
-    assert rt[0][1] == [[("a", "b")]]
-    assert rt[0][2] is not bnd and rt[0][2].lower == bnd.lower
-    assert rt[0][3] == "name"
-
-
-# ---- annotate tests ----
-
-class AnnRule:
-    def __init__(self, fn, bnd):
-        self._fn = fn
-        self._bnd = bnd
-    def get_annotation_function(self): return self._fn
-    def get_bnd(self): return self._bnd
-
-
-def test_annotate_returns_bounds_when_no_function():
-    bnd = _Interval(0.2, 0.3)
-    rule = AnnRule("", bnd)
-    lo, up = annotate([], rule, [], [])
-    assert (lo, up) == (0.2, 0.3)
-
-
-def test_annotate_calls_named_function():
-    bnd = _Interval(0, 1)
-    rule = AnnRule("foo", bnd)
-    def foo(ann, wts):
-        return (len(ann), len(wts))
-    out = annotate([foo], rule, [1, 2], [3])
-    assert out == (2, 1)
-
-
-# ---- float/str conversion helper tests ----
-
-def test_float_to_str_and_str_to_int():
-    assert float_to_str(12.345) == "12.345"
-    assert float_to_str(3.0) == "3.0"
-    assert str_to_int("123") == 123
-    assert str_to_int("-45") == -45
-
-
-@pytest.mark.parametrize(
-    "s,expected",
-    [("3.14", 3.14), ("42", 42.0), ("-2.5", -2.5)],
-)
-def test_str_to_float_variants(s, expected):
-    assert math.isclose(str_to_float(s), expected)
 
 
 def test_ground_rule_edge_recheck_failure_skips_body(monkeypatch):
